mxs: adapt default package list for Duckbill devices
Replace outdated package names, order the remaining.

firewall4 is added again, since it is enabled by default.
The device is not a router by default, so this package
makes no sense in the default config. User can enable it
by theirself, or it will be automatically pulled via
dependency, e.g. luci-app-firewall.

Signed-off-by: Michael Heimpold <mhei@heimpold.de>